# openapi.yaml
openapi: 3.0.3
info:
  title: PayFlow API
  version: "1.0.0"
  description: |
    Payment orchestration API for merchants. A merchant creates a Payment (starts
    `pending`), PayFlow asynchronously initiates it with the payment provider, and the
    Payment moves to `paid`/`failed` once the provider's webhook confirms the outcome -
    poll `GET /payments/{id}` (or, later, a merchant webhook) to observe the change.
    Refunds (partial or full) can be issued against a `paid`/`partially_refunded`
    payment; `sum(refunds) <= payment.amount` is enforced even under concurrent
    requests.

    All amounts are integers in the currency's smallest unit (e.g. haléře for CZK) -
    never a float.
servers:
  - url: http://127.0.0.1:8000/api
    description: Local development
security:
  - ApiKeyAuth: []
paths:
...
  /refunds/{refund}:
    get:
      summary: Get a single refund
      description: Same 404-not-403 reasoning as GET /payments/{payment}.
      tags: [Refunds]
      parameters:
        - name: refund
          in: path
          required: true
          schema: { type: string, format: ulid }
      responses:
        '200':
          description: The refund
          content:
            application/json:
              schema: { $ref: '#/components/schemas/RefundResource' }
        '401': { $ref: '#/components/responses/Unauthenticated' }
        '404': { $ref: '#/components/responses/NotFound' }
components:
  securitySchemes:
    ApiKeyAuth:
      type: http
      scheme: bearer
      description: |
        Merchant API key, issued out-of-band (no public self-registration endpoint
        exists yet). Send as `Authorization: Bearer <api_key>`.
...
  responses:
    Unauthenticated:
      description: Missing, invalid, or deactivated API key
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
          example: { error: { code: unauthenticated, message: "Invalid or missing API key." } }
    NotFound:
      description: Resource does not exist, or belongs to a different merchant
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
          example: { error: { code: not_found, message: "The requested resource was not found." } }
...
  schemas:
...
    Refund:
      type: object
      required: [id, payment_id, amount, idempotency_key, created_at]
      properties:
        id: { type: string, format: ulid }
        payment_id: { type: string, format: ulid }
        amount: { type: integer, example: 3000 }
        idempotency_key: { type: string }
        created_at: { type: string, format: date-time }
    RefundResource:
      type: object
      properties:
        data: { $ref: '#/components/schemas/Refund' }
...
    Error:
      type: object
      required: [error]
      properties:
        error:
          type: object
          required: [code, message]
          properties:
            code: { type: string, example: validation_failed }
            message: { type: string }
            details:
              type: object
              nullable: true
              additionalProperties: true
              description: Present on validation_failed - field name to array of messages.